rename chunk size config
